fix(chat-server): align room types and const locking

+++ exec/chat/chat-server/chat-server/User.h
@@ -15,7 +15,7 @@ public:
         std::shared_ptr<Session> session,
         uint64_t userId,
         std::string_view username,
-        uint64_t roomId);
+        uint32_t roomId);
 
     void Send(const flatbuffers::FlatBufferBuilder& builder) const;
 
@@ -25,11 +25,11 @@ public:
 
     [[nodiscard]] uint64_t GetId() const { return _id; }
     [[nodiscard]] std::string_view GetUsername() const { return _userName; }
-    [[nodiscard]] uint64_t GetRoomId() const { return _roomId; }
+    [[nodiscard]] uint32_t GetRoomId() const { return _roomId; }
 
 private:
     std::shared_ptr<Session> _session;
     uint64_t _id{0};
     std::string _userName;
-    uint64_t _roomId{0};
+    uint32_t _roomId{0};
 };
